[Dropbox] Dropbox Setup
Ok things are good to go on my end. I'll get you set up soon. Basically
right now there is a secret folder that must be placed into Source Art -
it's your Dropbox folder.

Git is going to ignore that and EVERYTHING in it, meaning you can
reference things in your project but we won't bog down commits with
music and art. It's on you to sync things from Dropbox. Let's test this
out and see how we like it!