[Light Cookies] Could Not Attain Desired Effect
I made a Light Cookie to try to even out the Pointlight's... light. But
it didn't work. The areas close to the light are still blown out, and
areas far away are very dim. Making the light more Intense doesn't fix
the problem because then when you get really close to an object the
whites get completely blown out.