[Light Shadows] Could Not Attain Desired Effect
This isn't going to work right now. The problem is this: if 2
Directional Lights are shining on the scene, they add light. Then when
another light tries to cast a shadow, that shadow isn't strong enough to
hide things. And I can't set that color.

The color would be set in the Shader, but here's the problem - that
Shader is applied to objects, not Lights! So we can't really have it
both ways, because then all of the other lights in the game will affect
objects with that Shader in the same weird way.

We might have to rely on the Shadow Volumes Toolkit for this, or we
won't be able to have the Shadow mechanic work on any level with lights.