Multiple Blockers for a single object
Rather than allowing multiple blockers to each partially block an object
from light (and thus allow it to change), I implemented a new method for
detecting if a monolith was in shadow which will allow for the blockers
to be of any size (rather than simply 1x1x1).

This will allow us to avoid situations where multiple blockers are
needed to put an object entirely in shadow.

Additionally, I added the ability for doorways to become closed (and
thus impassable).

See GitHub issue #14.