Point Light Shader Attenuation
Smoothed out the attenuation on the point lights, so that they have a
very crisp outer edge (not too crisp, just because it looks weird).

In order to update the shader, I had to switch our light rendering mode
from Forward to Deferred Legacy. This simply means that we can't use
some of the newer features of Unity 5, most importantly, the Standard
shader. I have updated the 'FlatShade' material to use a different
shader (Mobile/Diffuse), so you just have to apply that material (or any
other material not using the Standard shader) to any objects in your
scene. If this ends up being an issue, we can always look into other
solutions.

You can check out the Monolith scene for a great example of how it looks
now. Suck it, Unity!