[Sound Machine] Rough Draft of the Sound Machine
I got a basic Sound Machine going and also a scene for it. The scene
isn't really a level you play, it's just so you can listen to the sound.
Essentially there's a list of sounds and it picks one at random to play.

(The water sound is not part of the Sound Machine yet)

In the future, we'll be able to mix and match for each level in the game
so you can find the perfect blend of sounds. You never have full control
over the Sound Machine, but you can guide it in the right direction to
fit the theme.