Torch Interactions and Monolith State Rotation
Torches are now more user-friendly. If you attach a light to an object
with a torch script on it, the script will inherit the range, intensity,
and non-white color of that light.

Updated monolith state rotation - they can now loop or 'bounce' back and
forth. You can also specify that they should stop once they reach the
end.

This update was not made to portal monoliths - I updated them in a way
that more accurately shows how I think they should work. It's hard to
explain, so you should take a look.

See GitHub issue #34.