Make Torches Nicer and Fix Monoliths
Added the ability to change the rate of the flicker to the torches.

Added a 'block' object, which will cause the node underneath it to
become enabled/disabled when the object becomes active/inactive. Simply
attach it to any children of monoliths that should prevent the player
from moving to that position.

See GitHub issue #33.