[Sound Machine] Stream Implementation
* * Audio has not been mastered yet, do not listen on full volume with
headphones in! * *

Sound Machines are now organized further into Streams. Streams are
objects that hold Audio Clips and have some mixing values on them like
Calm Time (previously Chaos) that determine how frequently to play
sounds.

Listen to the game in SoundMachine.unity for a sample. Feel free to mess
with the SoundMachine on your own, or create one in your scene. It's
just a GameObject with SoundMachine on it and the values filled in with
AudioClips and numbers.