[Textures] Redone Floor and Wall Textures
I was messing around with textures and I remade the floor I liked in
Photoshop, along with some walls with ancient writing on them. I guess
I'm going for an "ancient ruins" kind of thing here? Unless that's too
cliche...

The writing on the wall can glow too if you set the Emission to greater
than 0 (in the standard shader). The material is m_Wall_Bot_Ruin.