Camera Movement and Shaking
Added camera movement and shaking. The fourth camera mode
(ACCEL_TO_PLAYER_MODE) is still a little glitchy.

I also added the camera shake to take effect whenever the player clicks.
That (probably) won't stay in the final version, but I liked how it
looked, so I wanted you to be able to see it. You can get rid of it (or
mess around with the parameters) by going to ClickDetector.cs.

Also got rid of the frame-long pause that the player used to do when he
reached a node. See GitHub issue #35.