First portion of first press release level
Made a first version of the first portion of the first press release
level. There are some questions that we're going to have to work through
in terms of visual design for this level.

Also, it doesn't look like 3D text is affected by shadows (or any
physics; it's always visible, even if it's behind something). So we'll
have to come up with another way to have disappearing text.