[Aesthetics] More Sounds, Particle Effects
I put the finishing touches on the Forest Sound Machine, the ambiance
sounds good now. Had to resort to using a fairly long clip of frogs (2
min) which I want to avoid in the future.

Then I worked on PFXManager, a script that implements Triggerable so
buttons can shut them off along with lights.