[Cutscene] Fighting Blender
This doesn't quite count as progress, because I fought with Blender this
entire time. I might actually switch over to 3DS Max... it's more
expensive, but I got a 3 year free trial today because of my Student
email (shhh) and they won't track us down unless we have the money to
pay anyway, so it should be alright.

Blender just has a crazy high learning curve with a really crappy
support infrastructure. Even when I get things to work, it's
unsatisfying because I never understand why.

This time, with 3DS Max, I'm going to write down every step involved in
the importing / exporting process. There will be a bit of relearning to
do, so this may slow down our progress a bit.