[Aesthetics] TREE!
Okay, the Tree is in a good place now so this task is done. It
highlighted an interesting problem - I suck at modeling "organic"
things, things that have natural curves and don't have right angles or
straight edges. Hopefully I'll find a way for the art style to display
stuff like trees in the future. I think I just need to go simpler...