[FX Manager] Non-Generic FX Manager
The current implementation of the FX Manager is half-and-half: some cues
are baked in with variables, whereas others use a new data structure
that is currently a work-in-progress.

The FX Manager is used to play sounds, trigger animations, swap
materials, and change lights during a cutscene. It is time-based and
suited to be in lockstep with the animation. This script is necessary
because I can't import sounds from 3DS Max.