[Cutscene] FX Manager, Cheats, and Cues
- The FX Manager has been tested. The tests are still in the cutscene
Scene but they're disabled. Activate them if you want to view the tests.
- I made it so you can set the reverb of a sound in the Cue. There's a
bunch of great presets that Unity provides. I used Cave for most of the
ones in this cutscene.
- I also added some cheats - the FX Manager allows you to change the
Time Scale of the game by pressing Left Arrow, Down Arrow, and Right
Arrow. This is to make cutscene editing a bit easier.

Left: Set it to Slow Motion
Down: Set it back to Normal
Right: Set it to Super Speed

AFTER LUNCH: Finish putting cues in, add Volume control for SFX