[Cutscene] Volume, Cues
I made it so you can lower the Volume of a Cue's audio. The scale is
from 0 to 1.

Then I went and put all the recorded audio things in as Cues. So there's
a ton of them. Don't change the size of the array in Cue List!! That is,
don't shrink it. That should do it for this initial test cutscene.