Player/Walker Troubles
The player and the walker weren't playing nice. Now they are.

This was inadvertently caused by the fix to the player pausing slightly
when he reached a node before continuing.

See GitHub issue #51.

Also see GitHub issue #50. It was addressed in the last commit, but I
forgot to mention it because I was so confused...