[Murder Fountain] Animation, FX Manager, Debugging
While working on this Murder Cutscene, I needed to animate the
protagonist dropping his lantern as he lunges for his victim. So now the
FX Manager can hold a Cue for changing an object's parent. This is
helpful because 3DS Max can't communicate parenting information with
Unity.

Unfortunately a lot of time was spent debugging... things get iffy in
Unity when parenting is involved. I ended up using an incredibly
convoluted solution, but it looks good in the end and that's what's
important. Woo?