Pretty Player, Cutscene Input Manager
I made some minor Pretty Player changes.

Then I made a small Cutscene Input Manager that does literally
everything EXCEPT pause the cutscene. That's going to be tougher...
maybe I can freeze time and space (the timescale) but I'm worried that
if I do that, the game won't be able to accept input and it will be
frozen forever.

If I can't do that, I need to overhaul the entire FX Manager system,
because it uses the game time. And the game time is gonna keep running
even if the animation is paused, which means pausing for even a split
second will cause everything to desync!

It astounds me how complicated something like pausing really is.

OH ALSO this is mouse clicking and not touch controls...